Sentiment analysis performed per candidate.
Sentiment analysis is done per candidate after basic data tidying. The results for the election will be coming on August 6 after the vote on August 5th. To keep things easy to maintain this is where the project will stop adding complications and begin refining the ways that sentiment analysis is done.

NOTE: Begin looking for standardized sentiment analysis methodologies.