Narrowing down focus of article scraper
Candidates narrowed to just mayoral. There was a visible symetry on the last set of plots that I want to investigate about the councilmember races. I will need to hold that off for later to focus on mayoral results and data validation while the tallies are being done.

NOTE: Kept to mayoral only until after race is validated.