Create README.md
Created a README file to summarize the work being done in this project. There are more notes in the commit logs about where in the projects maturation the development is and there are plans to add city council to the scrapers and sentimental analysis being done overall. A basic README has bee constructed to help organization.

NOTE: Personal Project